@@ -1,16 +1,16 @@
 {
-  "name": "h5p-course-presentation-alpha",
+  "name": "h5p-interactive-board",
   "version": "1.24.35",
   "description": "A free HTML5- based presentation content type which allows users to add multiple choice, fill in the blanks, texts and other types of interactions to their presentations using only a web browser.",
   "private": true,
   "scripts": {
-    "build": "cross-env NODE_ENV='production' webpack -p",
-    "watch": "webpack --watch",
-    "validate-types": "tsc"
+    "lint": "eslint './src/**/*.js'",
+    "build": "webpack --mode=production",
+    "watch": "webpack --mode=development --watch"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/h5p/h5p-course-presentation.git"
+    "url": "git+https://github.com/ndlano/h5p-interactive-board.git"
   },
   "keywords": [
     "h5p",
@@ -18,27 +18,27 @@
   ],
   "author": "Joubel",
   "license": "MIT",
+  "type": "module",
   "devDependencies": {
-    "babel-cli": "^6.23.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.23.0",
-    "babel-preset-latest": "^6.22.0",
-    "cross-env": "^5.2.0",
-    "css-loader": "^0.28.11",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "^0.11.2",
-    "h5p-lib-controls": "0.0.13",
+    "@babel/cli": "^7.23.4",
+    "@babel/core": "^7.23.7",
+    "@babel/preset-env": "^7.23.8",
+    "babel-loader": "^9.1.3",
+    "css-loader": "^6.9.0",
+    "eslint": "^8.56.0",
+    "eslint-config-ndla-h5p": "github:ndlano/eslint-config-ndla-h5p",
+    "h5p-lib-controls": "0.0.15",
     "h5p-parent": "0.0.2",
-    "resolve-url-loader": "^2.3.1",
-    "sass": "^1.32.8",
-    "sass-loader": "^7.3.1",
-    "style-loader": "^0.23.1",
-    "typescript": "^4.2.4",
-    "webpack": "^3.12.0"
+    "mini-css-extract-plugin": "^2.7.7",
+    "regenerator-runtime": "^0.14.1",
+    "resolve-url-loader": "^5.0.0",
+    "sass": "^1.69.7",
+    "sass-loader": "^13.3.3",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
   },
   "dependencies": {
-    "focus-trap": "^6.3.0"
-  }
+    "focus-trap": "^7.5.4"
+  },
+  "browserslist": "> 0.5%, last 3 versions, Firefox ESR, not dead"
 }
