@@ -1,4 +1,4 @@
-import { jQuery as $ } from './globals';
+import { jQuery as $ } from './globals.js';
 
 /**
  * @class
@@ -6,7 +6,6 @@ import { jQuery as $ } from './globals';
 export default class SlideBackground {
   /**
    * Create a Slide specific background selector
-   *
    * @param {H5P.InteractiveBoard} cp Course Presentation instance
    */
   constructor(cp) {
@@ -31,6 +30,41 @@ export default class SlideBackground {
       },
       params,
     );
+
+    /**
+     * Set background of slide(s)
+     * @private
+     * @param {object} fillSettings Background color settings
+     * @param {object} imageSettings Image background settings
+     * @param {number} [index] Optional target slide index, otherwise all slides.
+     */
+    var setBackground = function (fillSettings, imageSettings, index) {
+      var $updateSlides = cp.$slidesWrapper
+        .children()
+        .filter(':not(.h5p-summary-slide)');
+
+      if (index !== undefined) {
+        $updateSlides = $updateSlides.eq(index);
+      }
+
+      if (fillSettings && fillSettings !== '') {
+        // Fill with background color
+        $updateSlides
+          .addClass('has-background')
+          .css('background-image', '')
+          .css('background-color', fillSettings);
+      }
+      else if (imageSettings && imageSettings.path) {
+        // Fill with image
+        $updateSlides
+          .addClass('has-background')
+          .css('background-color', '')
+          .css(
+            'background-image',
+            'url(' + H5P.getPath(imageSettings.path, cp.contentId) + ')',
+          );
+      }
+    };
 
     /**
      * Set global background
@@ -59,41 +93,6 @@ export default class SlideBackground {
           );
         }
       });
-    };
-
-    /**
-     * Set background of slide(s)
-     *
-     * @private
-     * @param {Object} fillSettings Background color settings
-     * @param {Object} imageSettings Image background settings
-     * @param {number} [index] Optional target slide index, otherwise all slides.
-     */
-    var setBackground = function (fillSettings, imageSettings, index) {
-      var $updateSlides = cp.$slidesWrapper
-        .children()
-        .filter(':not(.h5p-summary-slide)');
-
-      if (index !== undefined) {
-        $updateSlides = $updateSlides.eq(index);
-      }
-
-      if (fillSettings && fillSettings !== '') {
-        // Fill with background color
-        $updateSlides
-          .addClass('has-background')
-          .css('background-image', '')
-          .css('background-color', fillSettings);
-      } else if (imageSettings && imageSettings.path) {
-        // Fill with image
-        $updateSlides
-          .addClass('has-background')
-          .css('background-color', '')
-          .css(
-            'background-image',
-            'url(' + H5P.getPath(imageSettings.path, cp.contentId) + ')',
-          );
-      }
     };
 
     // Set backgrounds
